yicong luo

+++ src/main/java/GroceryItem.java
@@ -4,7 +4,7 @@ public class GroceryItem{
     private double price;
     private int quantity;
 
-    public GroceryItem(String name, int price, int quantity) {
+    public GroceryItem(String name, double price, int quantity) {
         this.name = name;
         this.price = price;
         this.quantity = quantity;
